+++ cmd/cmd_test.go
@@ -2,43 +2,50 @@ package cmd
 
 import (
 	"testing"
+
+	"bytes"
+	"fmt"
 )
 
-func TestParseArgs(t *testing.T) {
-	type Case struct {
-		args     []string
-		expected Param
-	}
+// tests: tsk
+func TestTskMain(t *testing.T) {
+	// an io.Writer implementation for tests
+	buffer := &bytes.Buffer{}
 
-	cases := []Case{
-		{
-			// right args
-			[]string{"add", "implement task deletion"},
-			Param{"add", "implement task deletion"},
-		},
-		{
-			// action and number as arg
-			[]string{"do", "1"},
-			Param{"do", "1"},
-		},
-		{
-			// no action, no arg
-			[]string{},
-			Param{"", ""},
-		},
-		{
-			// too many args
-			[]string{"skip", "too", "many", "args"},
-			Param{"", ""},
-		},
-	}
+	// act
+	Main([]string{}, buffer)
 
-	for _, c := range cases {
-		got := ParseArgs(c.args)
-		if got != c.expected {
-			t.Errorf("FAIL: parseFlags(%v) == %v, expected %v", c.args, got, c.expected)
-		} else {
-			t.Logf("SUCCESS: parseFlags(%v) == %v, expected %v", c.args, got, c.expected)
-		}
+	// result
+	got := buffer.String()
+
+	// want
+	want := "No task, good news!\n"
+
+	// check
+	if got != want {
+		t.Errorf("got %q, want %q", got, want)
+	}
+}
+
+// tests: tsk add 'Implement task deletion'
+func TestTskMainAdd(t *testing.T) {
+	// command line arguments, got from os.Args
+	args := []string{"tsk", "add", "Implement task deletion"}
+
+	// an io.Writer implementation for tests
+	buffer := &bytes.Buffer{}
+
+	// act, in: args, out: buffer
+	Main(args, buffer)
+
+	// result
+	got := buffer.String()
+
+	// want
+	want := fmt.Sprintf("Added: 1. %s", args[2])
+
+	// check output
+	if got != want {
+		t.Errorf("got %q, want %q", got, want)
 	}
 }
