+++ main.go
@@ -1,6 +1,12 @@
 package main
 
+import (
+	"os"
+
+	"github.com/fitiavana07/tsk/cmd"
+)
+
 // main is the entrypoint in any go program
 func main() {
-	TskMain()
+	cmd.Main(os.Stdout)
 }
